+++ Kmse.Core/Cartridge/IMasterSystemCartridge.cs
@@ -5,5 +5,5 @@ public interface IMasterSystemCartridge
     Task<bool> LoadRomFromFile(string filename, CancellationToken cancellationToken);
     bool HasDumpHeader { get; }
     int Length { get; }
-    byte this[ushort address] { get; }
+    byte this[int address] { get; }
 }
